% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apg.R
\name{apg}
\alias{apg}
\title{Automobile Power Generation}
\usage{
apg(m, v, sa, crolling = 0.015, g = 9.8, p_air = 1.2, c_drag = 0.3)
}
\arguments{
\item{m:}{the mass of the vehicle in kg}

\item{v:}{the vehicle speed in m/s}

\item{sa:}{the surface area of the vehicle in m^2}

\item{crolling:}{the rolling resistive coefficient (default = 0.015)}

\item{c_drag:}{the aerodynamic resistive coefficient (default = 0.3)}

\item{g:}{the acceleration due to gravity (default = 9.8 m/s^2)}

\item{p_air:}{the density of air (default = 1.2 kg/m^3)}
}
\value{
power (W/s)
}
\description{
This function computes power generated from a
vehicle based on the following parameters:
}
\examples{
apg(m=1200, v = 40, sa = 10)
}
\author{
Nick Mc
}
