\name{clockstarG-package}
\alias{clockstarG-package}
\alias{clockstarG}
\docType{package}
\title{
clockstarG: Selecting the number of molecular-clocks for genomic data sets
}
\description{
clockstarG is an extension of ClockstaR, with capabilities to analyse very large genomic data sets.
}
\details{
\tabular{ll}{
Package: \tab clockstarG\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-31\cr
License: \tab GNU\cr
}
}
\author{
Sebastian Duchene
Maintainer: Sebastian Duchene <sebastian.duchene@sydney.edu.au>
}
\references{
Duchene, S, Molak, M., and Ho, SYW. "ClockstaR: choosing the number of relaxed-clock models in molecular phylogenetic analysis." Bioinformatics (2013): btt665.
}
\keyword{ molecular-clock }

