% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_yield.R
\name{crop_yield}
\alias{crop_yield}
\title{Crop Yields}
\usage{
crop_yield(fert, tp)
}
\arguments{
\item{fert}{the amount of fertilizer}

\item{tp}{the amount of total precipitation}
}
\value{
total yield (units unknown lol)
}
\description{
This function computes the average crop yield
based on the amount of fertilizer and precipitation
}
\author{
Nick Mc
}
