% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\name{fish_counts}
\alias{fish_counts}
\title{Fish function}
\usage{
fish_counts(fish)
}
\arguments{
\item{fish}{is a vector/txt file with list of fish names}
}
\value{
list of common fish, rarest fish, and total count
}
\description{
This function returns the most common, least common, and
total number of fish based on an input vector of fish names
}
\author{
Nick Mc
}
