% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_data.R
\docType{data}
\name{fish_data}
\alias{fish_data}
\title{Fish data}
\format{
A dataframe with 300 rows and 1 variable
\itemize{
\item fish (species)}
}
\source{
\url{https://naomitague.github.io/ESM_262_Programming/lectures/Class10_packages2.html#(5)}
}
\usage{
fish_data
}
\description{
Data of fish names that were caught of something.
}
\keyword{datasets}
