% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stocks.R
\name{ADS}
\alias{ADS}
\alias{GDAXI}
\alias{DOW}
\alias{load_all_stocks_DE}
\alias{load_all_stocks_US}
\alias{COMPONENTS_DE}
\alias{COMPONENTS_US}
\alias{global_controls_test_DE}
\alias{global_controls_test_US}
\title{Functions to load stock data}
\usage{
ADS()

GDAXI()

DOW()

load_all_stocks_DE()

load_all_stocks_US()

COMPONENTS_DE()

COMPONENTS_US()

global_controls_test_DE()

global_controls_test_US()
}
\description{
Functions to load stock data
}
