% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps_conv.R
\name{mps_conv}
\alias{mps_conv}
\title{Speed conversion}
\usage{
mps_conv(kph)
}
\arguments{
\item{kph}{The input speed in kilometers per hour}
}
\value{
mps The output speed in meters per second
}
\description{
Converts speed from kilometers per hour
to meters per second.
}
\author{
Nick Mc
}
