% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npv.R
\name{npv}
\alias{npv}
\title{Calculating Net Present Value}
\usage{
npv(r, i, t)
}
\arguments{
\item{r}{is net cash flow at time t}

\item{i}{is discount rate}

\item{t}{is the time of cash flow}
}
\value{
npv
}
\description{
Calculates net present value
}
\author{
Nick Mc
}
