% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFM_Function.R
\name{RFM_Function}
\alias{RFM_Function}
\title{Calculate RFM}
\usage{
RFM_Function(data, WeightRecency = 1/3, WeightFrequency = 1/3,
  WeightMonetary = 1/3)
}
\arguments{
\item{data}{- A data.table containing the transaction records.}

\item{WeightRecency}{- The weight for recency.}

\item{WeightFrequency}{- The weight for frequency.}

\item{WeightMonetary}{- The weight for monetary.}
}
\value{
Returns the score and group
}
\description{
Calculate the weighted RFM Score: recency, frequency, monetary value for every customer
}
\details{
\code{data} contains the transactional data
}
