% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\name{sliderinput_dates}
\alias{sliderinput_dates}
\alias{plot_stocks_DE}
\alias{hover_info_DE}
\alias{selectize_corona}
\alias{sliderinput_dates_corona}
\alias{plot_corona}
\alias{hover_info_corona}
\alias{selectize_Stocks_reg}
\alias{parameter_tabsi}
\alias{parameter_tabs}
\alias{tabs_custom}
\alias{parameter_tabsi_var}
\alias{parameter_tabs_var}
\alias{tabs_custom_var}
\alias{parameter_tabsi_xgb}
\alias{parameter_tabs_xgb}
\alias{tabs_custom_xgb}
\alias{numeric_features}
\title{UI Elements}
\usage{
sliderinput_dates()

plot_stocks_DE()

hover_info_DE()

selectize_corona()

sliderinput_dates_corona()

plot_corona()

hover_info_corona()

selectize_Stocks_reg()

parameter_tabsi()

parameter_tabs()

tabs_custom()

parameter_tabsi_var()

parameter_tabs_var()

tabs_custom_var()

parameter_tabsi_xgb()

parameter_tabs_xgb()

tabs_custom_xgb()

numeric_features()
}
\description{
UI Elements
}
