% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_model.R
\name{model_xgb}
\alias{model_xgb}
\alias{model_xgb_custom}
\alias{model_xgb_hyp}
\title{Plot Output Functions}
\usage{
model_xgb(res)

model_xgb_custom(
  res,
  mtry,
  trees,
  min_n,
  tree_depth,
  learn_rate,
  loss_reduction,
  sample_size
)

model_xgb_hyp(res, trees_hyp, grid_size)
}
\description{
Plot Output Functions
}
